Refactor partitioning back into a single command
