+++ cdip_admin/core/db/partitions/partition.py
@@ -13,17 +13,6 @@ class PARTITION_INTERVALS(Enum):
     MONTHLY = "monthly"
 
 
-class PartitionTableToolProtocol(Protocol):
-    def _create_parent_table(self) -> None:
-        raise NotImplementedError
-
-    def _partition_setup(self) -> None:
-        raise NotImplementedError
-
-    def _set_partition_schema_with_existing_tables(self) -> None:
-        raise NotImplementedError
-
-
 @dataclass
 class IndexData:
     name: str
@@ -58,7 +47,7 @@ class TableData:
     triggers: Optional[List[TriggerData]] = None
 
 
-class TablePartitionerBase(PartitionTableToolProtocol):
+class TablePartitionerBase:
     logger = logging.getLogger(__name__)
     steps_commands = [
         "_create_parent_table",
@@ -437,7 +426,7 @@ class TablePartitionerBase(PartitionTableToolProtocol):
         result = None
         with connection.cursor() as cursor:
             try:
-                self.logger.info(f"Executing SQL: {command} ...")
+                self.logger.debug(f"Executing SQL: {command} ...")
                 cursor.execute(command)
             except ProgrammingError as e:
                 self.logger.error(e)
@@ -446,308 +435,3 @@ class TablePartitionerBase(PartitionTableToolProtocol):
             if fetch:
                 result = cursor.fetchone()
         return result
-
-
-class ValuesListTablePartitioner(TablePartitionerBase):
-    def __init__(
-        self,
-        original_table_name: str,
-        partition_column: str,
-        partition_values: List[str],
-        subpartition_column: str,
-        table_data: TableData,
-    ) -> None:
-        super().__init__(
-            partition_column=partition_column,
-            original_table_name=original_table_name,
-            table_data=table_data,
-        )
-        self.subpartition_column = subpartition_column
-        self.partition_values = partition_values
-
-    def _partition_setup(self) -> None:
-        self.logger.info(
-            f"Applying partition setup for {self.partitioned_table_name} table using {self.template_table_name}..."
-        )
-        self.logger.info(
-            f"Creating partitions for values {self.partition_values}..."
-        )
-        # Create one partition for each value
-        for value in self.partition_values:
-            self.logger.info(
-                f"Creating partition for '{value}'..."
-            )
-            sql = f"""
-            CREATE TABLE {self.original_table_name}_{value} PARTITION OF public.{self.partitioned_table_name}
-            FOR VALUES IN ('{value}')
-            PARTITION BY RANGE ({self.subpartition_column});
-            """
-            self._execute_sql_command(command=sql)
-            self.logger.info(
-                f"Partition for '{value}' created."
-            )
-            self.logger.info(
-                f"Adding default sub-partition for '{value}'..."
-            )
-            sql = f"""
-            CREATE TABLE IF NOT EXISTS {self.original_table_name}_{value}_default
-            PARTITION OF {self.original_table_name}_{value} DEFAULT;
-            """
-            self._execute_sql_command(command=sql)
-            self.logger.info(
-                f"Default sub-partition for '{value}' created."
-            )
-        self.logger.info(
-            f"All partitions for values {self.partition_values} created."
-        )
-        self._set_current_step(step=3)
-        self.logger.info(
-            f"Partition setup for {self.partitioned_table_name} table using {self.template_table_name} is completed."
-        )
-
-    def _set_partition_schema_with_existing_tables(self) -> None:
-        self.logger.info(f"Setting Partition schema with existent tables...")
-        try:
-            self.logger.info(f"Locking tables...")
-            self._execute_sql_command(command="BEGIN;")
-            lock_tables_sql = f"""
-            LOCK TABLE public.{self.original_table_name} IN ACCESS EXCLUSIVE MODE;
-            LOCK TABLE public.{self.partitioned_table_name} IN ACCESS EXCLUSIVE MODE;
-            """
-            self._execute_sql_command(command=lock_tables_sql)
-            self.logger.info(f"Tables locked.")
-            self.logger.info(f"Renaming tables...")
-            # Original table is saved as backup
-            rename_tables_sql = f"""
-            ALTER TABLE public.{self.original_table_name}
-                RENAME TO {self.original_table_name}_original;
-
-            ALTER TABLE public.{self.partitioned_table_name}
-                RENAME TO {self.original_table_name};
-            """
-            self._execute_sql_command(command=rename_tables_sql)
-            self.logger.info(f"Tables renamed.")
-            self.logger.info(f"Creating default partition...")
-            create_default_part_sql = f"""
-            CREATE TABLE IF NOT EXISTS {self.original_table_name}_default
-            PARTITION OF {self.original_table_name} DEFAULT;
-            """
-            self._execute_sql_command(command=create_default_part_sql)
-            self.logger.info(f"Default partition created.")
-            self._execute_sql_command(command="COMMIT;")
-            self.logger.info(f"Setting Partition schema with existent tables completed.")
-
-        except Exception:
-            self._execute_sql_command(command="ROLLBACK;")
-            self.logger.exception("Failed to switch to partitioned table")
-            exit(1)
-
-        self._set_current_step(step=4)
-
-    def _process_data_partition(self) -> None:
-        self.logger.info("Moving existent data to partititons...")
-        self.logger.info(f"Locking table {self.original_table_name}...")
-        self._execute_sql_command(command="BEGIN;")
-        lock_table_sql = f"""LOCK TABLE public.{self.original_table_name} IN ACCESS EXCLUSIVE MODE;"""
-        self._execute_sql_command(command=lock_table_sql)
-        self.logger.info(f"Table {self.original_table_name} locked.")
-        for value in self.partition_values:
-            partition_name = f"{self.original_table_name}_{value}"
-            self.logger.info(f"Moving data of type '{value}' to partition {partition_name} ...")
-            # ToDo: Copy data in batches
-            migrate_sql = f"""
-            INSERT INTO {self.original_table_name}_{value}
-            SELECT * FROM public.{self.original_table_name}_original
-            WHERE log_type = '{value}';
-            """
-            self._execute_sql_command(command=migrate_sql)
-        self._execute_sql_command(command="COMMIT;")
-        self.logger.info("Moving existent data to partitions...completed")
-
-        self.logger.info("Running VACUUM ANALYZE...")
-        self._execute_sql_command(command=f"VACUUM ANALYZE public.{self.original_table_name};")
-        self._execute_sql_command(command="VACUUM;")
-        self.logger.info("VACUUM ANALYZE is completed.")
-
-        self.logger.info("Restoring triggers...")
-        for trigger in self.table_data.triggers if self.table_data.triggers else []:
-            self._drop_trigger(table_name=f"{self.original_table_name}_original", trigger_data=trigger)
-            self._create_trigger(table_name=self.original_table_name, trigger_data=trigger)
-        self.logger.info("Triggers restored")
-
-        self.logger.info(f"Creating unique index on PK {self.table_data.primary_key_columns}...")
-        pk_unique_idx_name = f"{'_'.join(self.table_data.primary_key_columns)}_unique_idx"
-        self._create_index(
-            table_name=self.original_table_name,
-            index_data=IndexData(name=pk_unique_idx_name, columns=self.table_data.primary_key_columns),
-            is_unique=True,
-        )
-        self.logger.info("Unique index on PK created.")
-
-        self.logger.info("Restoring unique constraints...")
-        for unique_constraint in self.table_data.unique_constraints if self.table_data.unique_constraints else []:
-            self._create_unique_constraint(table_name=self.original_table_name, constraint_data=unique_constraint)
-        self.logger.info("Unique constraints restored.")
-
-        self._set_current_step(step=6)
-
-
-class DateRangeTablePartitioner(TablePartitionerBase):
-
-    def __init__(
-        self,
-        original_table_name: str,
-        partition_column: str,
-        partition_interval: str,
-        table_data: TableData,
-        partition_start: str,
-        migrate_batch_size_per_interval: int,
-        partitions_in_the_future: int = 5,
-    ) -> None:
-        super().__init__(
-            partition_column=partition_column,
-            original_table_name=original_table_name,
-            table_data=table_data,
-        )
-        self.partition_interval = partition_interval
-        self.migrate_batch_size = migrate_batch_size_per_interval
-        self.partition_start = partition_start
-        self.partitions_in_the_future = partitions_in_the_future
-
-    def _partition_setup(self) -> None:
-        self.logger.info(
-            f"Applying partition setup for {self.partitioned_table_name} table using {self.template_table_name}..."
-        )
-        # ToDo: Ask why they are not using partman for creating partitions (past and future)
-        sql = f"""
-            SELECT partman.create_parent(
-               p_parent_table := 'public.{self.partitioned_table_name}',
-               p_control := '{self.partition_column}',
-               p_type := 'native',
-               p_start_partition := '{self.partition_start}',
-               p_interval := '{self.partition_interval}',
-               p_template_table := 'public.{self.template_table_name}',
-               p_premake := {self.partitions_in_the_future}
-               );
-        """
-        self._execute_sql_command(command=sql)
-
-        # Remove default partition created by partman.
-        # The original table will be used as default partition.
-        sql = f"DROP TABLE public.{self.partitioned_table_name}_default;"
-        self._execute_sql_command(command=sql)
-
-        drop_default_partitions_sql = f"""
-            DO $$
-            DECLARE
-                partition_table TEXT;
-            BEGIN
-                FOR partition_table IN
-                    SELECT partman.show_partitions('public.{self.partitioned_table_name}')
-                LOOP
-                    EXECUTE 'DROP TABLE IF EXISTS ' || REGEXP_REPLACE(partition_table::TEXT, '\\(.*?,(.*)\\)', '\\1');
-                END LOOP;
-            END $$;
-            """
-        self._execute_sql_command(command=drop_default_partitions_sql)
-
-        sql = f"""
-        UPDATE partman.part_config
-        SET parent_table = 'public.{self.original_table_name}'
-        WHERE parent_table = 'public.{self.partitioned_table_name}';
-        """
-        self._execute_sql_command(command=sql)
-        self._set_current_step(step=3)
-        self.logger.info(
-            f"Partition setup for {self.partitioned_table_name} table using {self.template_table_name} is completed."
-        )
-
-    def _set_partition_schema_with_existing_tables(self) -> None:
-        self.logger.info(f"Setting Partition schema with existent tables...")
-        try:
-            self.logger.info(f"Locking tables...")
-            self._execute_sql_command(command="BEGIN;")
-            lock_tables_sql = f"""
-            LOCK TABLE public.{self.original_table_name} IN ACCESS EXCLUSIVE MODE;
-            LOCK TABLE public.{self.partitioned_table_name} IN ACCESS EXCLUSIVE MODE;
-            """
-            self._execute_sql_command(command=lock_tables_sql)
-            self.logger.info(f"Tables locked.")
-            self.logger.info(f"Renaming tables...")
-            # Original table is turned into the default partition
-            rename_tables_sql = f"""
-            ALTER TABLE public.{self.original_table_name}
-                RENAME TO {self.original_table_name}_default;
-
-            ALTER TABLE public.{self.partitioned_table_name}
-                RENAME TO {self.original_table_name};
-            """
-            self._execute_sql_command(command=rename_tables_sql)
-            self.logger.info(f"Tables renamed.")
-            # ToDo: this fails if the table is a partition already
-            # self.logger.info(f"Attaching default partition...")
-            # attach_sql = f"""
-            # ALTER TABLE public.{self.original_table_name}
-            # ATTACH PARTITION public.{self.original_table_name}_default DEFAULT;
-            # """
-            # self._execute_sql_command(command=attach_sql)
-            # self.logger.info(f"Default partition attached.")
-            self._execute_sql_command(command="COMMIT;")
-            self.logger.info(f"Setting Partition schema with existent tables completed.")
-
-        except Exception:
-            self._execute_sql_command(command="ROLLBACK;")
-            self.logger.exception("Failed to switch to partitioned table")
-            exit(1)
-
-        self._set_current_step(step=4)
-
-    def _process_data_partition(self) -> None:
-        self.logger.info("Moving existent data to partititons...")
-        # self._execute_sql_command(command="BEGIN;")
-        # self.logger.info(f"Copying data from backup to partitioned table {self.original_table_name} ...")
-        # migrate_sql = f"""
-        # INSERT INTO {self.original_table_name}
-        # SELECT * FROM public.{self.original_table_name}_original;
-        # """
-        # self._execute_sql_command(command=migrate_sql)
-        # self._execute_sql_command(command="COMMIT;")
-        # self.logger.info(f"Backup data copied to partitioned table {self.original_table_name}.")
-        self.logger.info(f"Processing data with partman.partition_data_proc in batches ...")
-        migrate_sql = f"""
-            CALL partman.partition_data_proc(
-            'public.{self.original_table_name}',
-            p_wait:= 2,
-            p_batch := {self.migrate_batch_size}
-            );
-        """
-        self.logger.info(f"partman.partition_data_proc completed.")
-        self._execute_sql_command(command=migrate_sql)
-        self.logger.info("Moving existent data to partititons...completed")
-
-        self.logger.info("Running VACUUM ANALYZE...")
-        self._execute_sql_command(command=f"VACUUM ANALYZE public.{self.original_table_name};")
-        self._execute_sql_command(command="VACUUM;")
-        self.logger.info("VACUUM ANALYZE is completed.")
-
-        self.logger.info("Restoring triggers...")
-        for trigger in self.table_data.triggers if self.table_data.triggers else []:
-            self._drop_trigger(table_name=f"{self.original_table_name}_default", trigger_data=trigger)
-            self._create_trigger(table_name=self.original_table_name, trigger_data=trigger)
-        self.logger.info("Triggers restored")
-
-        self.logger.info("Creating unique index on PK...")
-        pk_unique_idx_name = f"{'_'.join(self.table_data.primary_key_columns)}_unique_idx"
-        self._create_index(
-            table_name=self.original_table_name,
-            index_data=IndexData(name=pk_unique_idx_name, columns=self.table_data.primary_key_columns),
-            is_unique=True,
-        )
-        self.logger.info("Unique index on PK created.")
-
-        self.logger.info("Restoring unique constraints...")
-        for unique_constraint in self.table_data.unique_constraints if self.table_data.unique_constraints else []:
-            self._create_unique_constraint(table_name=self.original_table_name, constraint_data=unique_constraint)
-        self.logger.info("Unique constraints restored.")
-
-        self._set_current_step(step=5)
